Keep the changes, I'll integrate/port what i can from the example. I am just more comfortable making stuff like this.

+++ 2D Shooter/src/com/entity/GameObject.java
@@ -5,11 +5,13 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-    protected float x, y;
+    protected int x, y;
     protected ID id;
-    protected float velX, velY;
+    protected int velX, velY;
 
-    public GameObject(float x, float y, ID id){
+
+
+    public GameObject(int x, int y, ID id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -28,11 +30,11 @@ public abstract class GameObject {
         this.y = y;
     }
 
-    public float getX(){
+    public int getX(){
         return x;
     }
 
-    public float getY(){
+    public int getY(){
         return y;
     }
 
@@ -44,19 +46,19 @@ public abstract class GameObject {
         return id;
     }
 
-    public void setVelX(float velX){
+    public void setVelX(int velX){
         this.velX = velX;
     }
 
-    public void setVelY(float velY){
+    public void setVelY(int velY){
         this.velY = velY;
     }
 
-    public float getVelX(){
+    public int getVelX(){
         return velX;
     }
 
-    public float getVelY(){
+    public int getVelY(){
         return velY;
     }
